Update Setup Command to update $namespace variable
As the Setup command runs a str_replace and ucwords on the inputted $namespace it's often different than what's input. The output later in the setup then can be misleading as in some cases it'll have changed during the process to strip spaces and make first words uppercase.
This patch updates the $namespace variable so when it's output later the user see what the actual project namespace is.